[fix] 홈에서 개봉 화면 넘어갈 때 간단한 주소 출력 (#101)
