+++ SpaceCapsule/SpaceCapsule/Scene/TabBar/Home/HomeViewModel.swift
@@ -10,7 +10,7 @@ import Foundation
 import RxCocoa
 import RxSwift
 
-final class HomeViewModel: BaseViewModel {
+final class HomeViewModel: BaseViewModel, CapsuleCellNeedable {
     weak var coordinator: HomeCoordinator?
     let disposeBag = DisposeBag()
 
@@ -60,7 +60,10 @@ final class HomeViewModel: BaseViewModel {
         input.tapCapsule
             .withUnretained(self)
             .subscribe(onNext: { owner, uuid in
-                owner.coordinator?.moveToCapsuleAccess(uuid: uuid)
+                guard let capsuleItemCell = owner.getCellItem(with: uuid) else {
+                    return
+                }
+                owner.coordinator?.moveToCapsuleAccess(with: capsuleItemCell)
             })
             .disposed(by: disposeBag)
